Изменил источник новостей Пофиксил баг

+++ AssistantJula_bot/Model/Weather/Wind.cs
@@ -8,6 +8,6 @@
 		/// <summary>
 		/// Скорость ветра
 		/// </summary>
-		public int speed { get; set; }
+		public float speed { get; set; }
 	}
 }
